make axis conversion available to LumiTransientSpec
