+++ lumispy/signals/luminescence_spectrum.py
@@ -23,22 +23,13 @@ from warnings import warn
 
 from hyperspy.signals import Signal1D
 from hyperspy._signals.lazy import LazySignal
-from hyperspy.axes import DataAxis
 from traits.api import Undefined
 
 from lumispy.signals.common_luminescence import CommonLumi
-from lumispy import nm2invcm, to_array, savetxt
+from lumispy import to_array, savetxt
+from lumispy.utils import solve_grating_equation
 from lumispy.utils.axes import GRATING_EQUATION_DOCSTRING_PARAMETERS
 from lumispy.utils.signals import com
-from lumispy.utils import (
-    axis2eV,
-    data2eV,
-    var2eV,
-    axis2invcm,
-    data2invcm,
-    var2invcm,
-    solve_grating_equation,
-)
 from lumispy.utils.io import (
     SAVETXT_DOCSTRING,
     SAVETXT_PARAMETERS,
@@ -55,295 +46,6 @@ class LumiSpectrum(Signal1D, CommonLumi):
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-
-    def _reset_variance_linear_model(self):
-        """Resets the variance linear model parameters to their default values,
-        as they are not applicable any longer after a Jacobian transformation.
-        """
-        if (
-            (
-                self.metadata.has_item(
-                    "Signal.Noise_properties.Variance_linear_model.gain_factor"
-                )
-                and self.metadata.Signal.Noise_properties.Variance_linear_model.gain_factor
-                != 1
-            )
-            or (
-                self.metadata.has_item(
-                    "Signal.Noise_properties.Variance_linear_model.gain_offset"
-                )
-                and self.metadata.Signal.Noise_properties.Variance_linear_model.gain_offset
-                != 0
-            )
-            or (
-                self.metadata.has_item(
-                    "Signal.Noise_properties.Variance_linear_model.correlation_factor"
-                )
-                and self.metadata.Signal.Noise_properties.Variance_linear_model.correlation_factor
-                != 1
-            )
-        ):
-            self.metadata.Signal.Noise_properties.Variance_linear_model.gain_factor = 1
-            self.metadata.Signal.Noise_properties.Variance_linear_model.gain_offset = 0
-            self.metadata.Signal.Noise_properties.Variance_linear_model.correlation_factor = (
-                1
-            )
-            warn(
-                "Following the Jacobian transformation, the parameters of the "
-                "`Variance_linear_model` are reset to their default values "
-                "(gain_factor=1, gain_offset=0, correlation_factor=1).",
-                UserWarning,
-            )
-
-    def _convert_data(self, newaxis, factor, inplace, jacobian, data2, var2):
-        """Utility function to perform the data and variance conversion for
-        signal unit transformations.
-        """
-        # convert data
-        if jacobian:
-            s2data = np.copy(
-                data2(
-                    self.isig[::-1].data,
-                    factor,
-                    newaxis.axis,
-                    self.axes_manager.signal_axes[0],
-                )
-            )
-        else:
-            s2data = np.copy(self.isig[::-1].data)
-
-        # inplace conversion
-        if inplace:
-            s2 = self
-            s2.data = s2data
-        # create new signal object with correct data, axes, metadata
-        else:
-            s2 = self._deepcopy_with_new_data(
-                s2data,
-                copy_variance=(not jacobian),
-                copy_learning_results=False,
-            )
-        # convert axis
-        oldaxis = self.axes_manager.signal_axes[0]
-        s2.axes_manager.set_axis(
-            newaxis,
-            self.axes_manager.signal_axes[0].index_in_axes_manager,
-        )
-        # convert variance
-        if self.metadata.has_item("Signal.Noise_properties.variance"):
-            var = self.get_noise_variance()
-            if jacobian:
-                # if variance is a numeric value, cast into signal object
-                if isinstance(var, (float, int)):
-                    var = self._deepcopy_with_new_data(
-                        np.ones(self.data.shape) * var,
-                        copy_variance=False,
-                        copy_learning_results=False,
-                    )
-                s2var = s2._deepcopy_with_new_data(
-                    var2(
-                        var.isig[::-1].data,
-                        factor,
-                        newaxis.axis,
-                        oldaxis,
-                    ),
-                    copy_variance=False,
-                    copy_learning_results=False,
-                )
-                s2.set_noise_variance(s2var)
-                s2._reset_variance_linear_model()
-            else:
-                # variance left unchanged, if it is a number and Jacobian not performed
-                if isinstance(var, (float, int)):
-                    if not inplace:
-                        s2.set_noise_variance(self.get_noise_variance())
-                else:
-                    s2.set_noise_variance(
-                        s2._deepcopy_with_new_data(
-                            var.isig[::-1].data,
-                            copy_variance=False,
-                            copy_learning_results=False,
-                        )
-                    )
-        if not inplace:
-            return s2
-        else:
-            return None
-
-    def to_eV(self, inplace=True, jacobian=True):
-        """Converts signal axis of 1D signal to non-linear energy axis (eV)
-        using wavelength dependent refractive index of air. Assumes wavelength
-        in units of nm unless the axis units are specifically set to µm.
-
-        The intensity is converted from counts/nm (counts/µm) to counts/meV by
-        doing a Jacobian transformation, see e.g. Mooney and Kambhampati, J.
-        Phys. Chem. Lett. 4, 3316 (2013), doi:10.1021/jz401508t, which ensures
-        that integrated signals are correct also in the energy domain. If the
-        variance of the signal is known, i.e.
-        `metadata.Signal.Noise_properties.variance` is a signal representing
-        the variance, a squared renormalization of the variance is performed.
-        Note that if the variance is a number (not a signal instance), it is
-        converted to a signal if the Jacobian transformation is performed
-
-        Parameters
-        ----------
-        inplace : boolean
-            If `False`, a new signal object is created and returned. Otherwise
-            (default) the operation is performed on the existing signal object.
-        jacobian : boolean
-            The default is to do the Jacobian transformation (recommended at
-            least for luminescence signals), but the transformation can be
-            suppressed by setting this option to `False`.
-
-        Examples
-        --------
-        >>> import numpy as np
-        >>> from lumispy import LumiSpectrum
-        >>> S1 = LumiSpectrum(np.ones(20), DataAxis(axis = np.arange(200,400,10)), ))
-        >>> S1.to_eV()
-        """
-
-        evaxis, factor = axis2eV(self.axes_manager.signal_axes[0])
-
-        return self._convert_data(evaxis, factor, inplace, jacobian, data2eV, var2eV)
-
-    TO_INVCM_DOCSTRING = """
-        The intensity is converted from counts/nm (counts/µm) to counts/cm^-1
-        by doing a Jacobian transformation, see e.g. Mooney and Kambhampati,
-        J. Phys. Chem. Lett. 4, 3316 (2013), doi:10.1021/jz401508t, which
-        ensures that integrated signals are correct also in the wavenumber
-        domain. If the variance of the signal is known, i.e.
-        `metadata.Signal.Noise_properties.variance` is a signal representing the
-        variance, a squared renormalization of the variance is performed.
-        Note that if the variance is a number (not a signal instance), it is
-        converted to a signal if the Jacobian transformation is performed
-
-        Parameters
-        ----------
-        inplace : boolean
-            If `False`, a new signal object is created and returned. Otherwise
-            (default) the operation is performed on the existing signal object.
-        """
-
-    TO_INVCM_EXAMPLE = """
-        Examples
-        --------
-        >>> import numpy as np
-        >>> from lumispy import LumiSpectrum
-        >>> S1 = LumiSpectrum(np.ones(20), DataAxis(axis = np.arange(200,400,10)), ))
-        >>> S1.to_invcm()
-        """
-
-    def to_invcm(self, inplace=True, jacobian=True):
-        """Converts signal axis of 1D signal to non-linear wavenumber axis
-        (cm^-1). Assumes wavelength in units of nm unless the axis units are
-        specifically set to µm.
-        %s
-        jacobian : boolean
-            The default is to do the Jacobian transformation (recommended at
-            least for luminescence signals), but the transformation can be
-            suppressed by setting this option to `False`.
-        %s
-        """
-
-        invcmaxis, factor = axis2invcm(self.axes_manager.signal_axes[0])
-
-        return self._convert_data(
-            invcmaxis, factor, inplace, jacobian, data2invcm, var2invcm
-        )
-
-    to_invcm.__doc__ %= (TO_INVCM_DOCSTRING, TO_INVCM_EXAMPLE)
-
-    TO_INVCMREL_EXAMPLE = """
-        Examples
-        --------
-        >>> import numpy as np
-        >>> from lumispy import LumiSpectrum
-        >>> S1 = LumiSpectrum(np.ones(20), DataAxis(axis = np.arange(200,400,10)), ))
-        >>> S1.to_invcm(laser=325)
-    """
-
-    def to_invcm_relative(self, laser=None, inplace=True, jacobian=False):
-        """Converts signal axis of 1D signal to non-linear wavenumber axis
-        (cm^-1) relative to the exciting laser wavelength (Raman/Stokes
-        shift). Assumes wavelength in units of nm unless the axis units are
-        specifically set to µm.
-        %s
-        laser: float or None
-            Laser wavelength in the same units as the signal axis. If None
-            (default), checks if it is stored in
-            `metadata.Acquisition_instrument.Laser.wavelength`.
-        jacobian : boolean
-            The default is not to do the Jacobian transformation for Raman
-            shifts, but the transformation can be activated by setting this
-            option to `True`.
-        %s
-        """
-
-        # check if laser wavelength is available
-        if laser is None:
-            if not self.metadata.has_item("Acquisition_instrument.Laser.wavelength"):
-                raise AttributeError(
-                    "Laser wavelength is neither given in the metadata nor passed"
-                    " to the function."
-                )
-            else:
-                laser = self.metadata.get_item(
-                    "Acquisition_instrument.Laser.wavelength"
-                )
-        # check if laser units make sense in respect to signal units
-        if (self.axes_manager.signal_axes[0].units == "µm" and laser > 10) or (
-            self.axes_manager.signal_axes[0].units == "nm" and laser < 100
-        ):
-            raise AttributeError(
-                "Laser wavelength units do not seem to match the signal units."
-            )
-
-        invcmaxis, factor = axis2invcm(self.axes_manager.signal_axes[0])
-
-        # convert to relative wavenumber scale
-        if self.axes_manager.signal_axes[0].units == "µm":
-            invcmlaser = nm2invcm(1000 * laser)
-        else:
-            invcmlaser = nm2invcm(laser)
-        absaxis = invcmaxis.axis[::-1]
-        invcmaxis.axis = invcmlaser - absaxis
-        invcmaxis.name = "Raman Shift"
-
-        # replace signal axis after conversion
-        # conversion (using absolute scale for Jacobian)
-        if inplace:
-            self.to_invcm(inplace=inplace, jacobian=jacobian)
-            s2 = self
-        else:
-            s2 = self.to_invcm(inplace=inplace, jacobian=jacobian)
-        # replace axis
-        s2.axes_manager.set_axis(
-            invcmaxis,
-            self.axes_manager.signal_axes[0].index_in_axes_manager,
-        )
-        s2.data = s2.isig[::-1].data
-        # replace variance axis
-        if s2.metadata.has_item("Signal.Noise_properties.variance") and not isinstance(
-            s2.get_noise_variance(), (float, int)
-        ):
-            s2.metadata.Signal.Noise_properties.variance.axes_manager.set_axis(
-                invcmaxis,
-                s2.axes_manager.signal_axes[0].index_in_axes_manager,
-            )
-            s2.metadata.Signal.Noise_properties.variance.data = (
-                s2.metadata.Signal.Noise_properties.variance.isig[::-1].data
-            )
-        if not inplace:
-            return s2
-
-    to_invcm_relative.__doc__ %= (
-        TO_INVCM_DOCSTRING,
-        TO_INVCMREL_EXAMPLE,
-    )
-
-    # Alias Method Name
-    to_raman_shift = to_invcm_relative
 
     def remove_background_from_file(self, background=None, inplace=False, **kwargs):
         """Subtract the background to the signal in all navigation axes. If no
